WebCryptoAPI/RSA-OAEP.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>RSA-OAEP</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="util.js"></script>
<script src="test-vectors.js"></script>
<div id=log></div>
<script>
"use strict";

promise_test(function() {
  var alg = {name: "RSA-OAEP", hash: "SHA-1"};

  return Promise.all([
    crypto.subtle.importKey("pkcs8", tv.rsaoaep.pkcs8, alg, false, ['decrypt']),
    crypto.subtle.importKey("spki", tv.rsaoaep.spki, alg, false, ['encrypt'])
      .then(function(pubKey) { return crypto.subtle.encrypt(alg, pubKey, tv.rsaoaep.data); })
  ]).then(function(values) {
    var privKey = values[0];
    var encrypted = values[1];
    return crypto.subtle.decrypt(alg, privKey, encrypted);
  }).then(function(x) {
    assert_true(util.memcmp(x, tv.rsaoaep.data));
  });
}, "RSA-OAEP encrypt/decrypt round-trip");

promise_test(function() {
  var alg = {
    name: "RSA-OAEP",
    hash: "SHA-256",
    modulusLength: 2048,
    publicExponent: new Uint8Array([0x01, 0x00, 0x01])
  };

  var data = crypto.getRandomValues(new Uint8Array(128));

  return crypto.subtle.generateKey(alg, false, ['encrypt', 'decrypt']).then(function(x) {
    return Promise.all([
      x.privateKey,
      crypto.subtle.encrypt(alg, x.publicKey, data)
    ]);
  }).then(function(values) {
    var privKey = values[0];
    var encrypted = values[1];
    return crypto.subtle.decrypt(alg, privKey, encrypted);
  }).then(function(x) {
    assert_true(util.memcmp(x, data));
  });
}, "RSA-OAEP key generation and encrypt/decrypt round-trip (SHA-256)");

promise_test(function() {
  var alg = {name: "RSA-OAEP", hash: "SHA-1"};
  return crypto.subtle.importKey("pkcs8", tv.rsaoaep.pkcs8, alg, false,
                                 ['decrypt']).then(function(x) {
    return crypto.subtle.decrypt(alg, x, tv.rsaoaep.result);
  }).then(function(x) {
    assert_true(util.memcmp(x, tv.rsaoaep.data));
  });
}, "RSA-OAEP decryption known answer");

promise_test(function(test) {
  var alg = {
    name: "RSA-OAEP",
    hash: "SHA-1",
    modulusLength: 2048,
    publicExponent: new Uint8Array([0x01, 0x00, 0x01])
  };

  return crypto.subtle.generateKey(alg, false, ['encrypt', 'decrypt']).then(function(x) {
    var pubKey = x.publicKey;
    return Promise.all([
      pubKey,
      crypto.subtle.encrypt(alg, pubKey, new Uint8Array(214)),
    ]);
  }).then(function(values) {
    var pubKey = values[0];
    return promise_rejects(test, null /*XXX "OperationError"*/,
      crypto.subtle.encrypt(alg, pubKey, new Uint8Array(215)));
  });
}, "RSA-OAEP input data length checks (2048-bit key)");

promise_test(function(test) {
  var alg = {name: "RSA-OAEP", hash: "SHA-123"};

  return promise_rejects(test, "DataError",
    crypto.subtle.importKey("pkcs8", tv.rsaoaep.pkcs8, alg, false, ['decrypt']));
}, "RSA-OAEP key import with invalid hash");

promise_test(function(test) {
  var alg = {
    name: "RSA-OAEP",
    hash: "SHA-256",
    modulusLength: 2048,
    publicExponent: new Uint8Array([0x01, 0x00, 0x01])
  };

  var data = crypto.getRandomValues(new Uint8Array(128));

  return crypto.subtle.generateKey(alg, false, ["encrypt", "decrypt"]).then(function(x) {
    return Promise.all([
      x.privateKey,
      crypto.subtle.encrypt("RSA-OAEP", x.publicKey, data),
    ]);
  }).then(function doDecrypt(values) {
    var privKey = values[0];
    var encrypted = values[1];
    return crypto.subtle.decrypt("RSA-OAEP", privKey, encrypted);
  }).then(function(x) {
    assert_true(util.memcmp(x, data));
  });
}, "Test that RSA-OAEP encrypt/decrypt accepts strings as AlgorithmIdentifiers");
</script>
